change minor css and syntax, optimize OrgLayout
